(housekeep) deleting unnecery files copied from org-roam
As I have re-architected the whole md-roam as a plug-in for org-roam,
I needed to delete the files copied from the original project.
It seems that they were used (as they were presumably compiled after the original version).
This perhaps caused the md-roam plug-in to terminate with an unexpected error.