(fix) duplicate titles
The duplicate titles issue is caused by the change introduced to org-roam--extract-titles.
Upstream change (feat): optionally use headline as title (#538).
If the function cannot find org titile (#+TITLE) then, it takes the first heading
(not necessarily the first line, as evident in my literature note with #+ROAM_REF.)

My fix this time is rather crude (need to come back to this); I'm overriding org-roam's
titles, including ALIASES, and add md-roam title (title:).

This means ALIASES are currently not supported. I don't use them. But it does not feel
very good as a long-term solution. In addtion, this also means you cannot use the first
line as the title (But I don't think I want this as a feature; file name works when
title is missing for markdown files).