(fix) Use upstream org-roam-title-sources variable
Upstream commit b47b76a (#625) refactors title extraction
function, introduces a new variable to add flexibility to how
the titles are identified and extracted. This fix takes
advantage of it, and adds a new function to extract the title of
a markdown file.

This commit also adds the optional FORCE argument to the advising
function to build-cache function to fix the problem of argument
mismatch.