fix #69 attachment to md files output a read-only buffer warning
This seems to be caused by the fact that markdown file links are read
and opened by `find-file-noselect`. This is necessary to get the file's
absolute file name to be inserted to the DB; however, the issue is if
the attached file is a read-only file such as PDF or image, each link
promts asking the user for an Y/N input -- this is no good when you have
multiple links over many files.

- File extension is checked for each link. Only links to files with
  the `md-roam-file-extension` will be insert to the DB.

- `find-file-noselect` is called with NOWARN to avoid warning in
   database udpate.

When a file link is to a non-existing markdown file, you seem to still
get a warning about `make-directory`. I am leaving this for now.